edge-surveillance-system/event-bus/requirements.txt
Flask==2.3.2
docker==6.1.3
PyYAML==6.0
requests==2.31.0